TestRestApi/product_catalog_api/views/api_root.py
from rest_framework.decorators import api_view
from rest_framework.response import Response
from rest_framework.reverse import reverse


@api_view(['GET'])
def api_root(request, format=None):
    return Response({
        'products': reverse('product-list', request=request, format=format),
        'product-register':
            reverse('productregister-list', request=request, format=format),
    })
