TestRestApi/requirements.txt
django>=2.1,<=2.1.99
djangorestframework
django-admin
psycopg2-binary
Pillow
django-filter